Docs versioning (#339)
Build different Bow docs versions based on git tags, making them available to consult on the website.

The different versions are read from the Git tags present in the project. This is done through a Ruby script (`gen_docs.rb`), which run in a Travis environment, and the script supports different options that can be set, e.g. the tags for which we want docs to be generated from.

This only sets the CI to build the different version microsites under a structure suitable to be used moving forward. The version selector is being habdled through another PR (#334)